+++ lib/lite/command/internals/executions.rb
@@ -60,7 +60,7 @@ module Lite
           around_execution { call }
           Utils.try(self, :on_success)
         rescue StandardError => e
-          @exception = e
+          @original_exception = e
           fault(e, ERROR, metadata) unless e.is_a?(Lite::Command::Fault)
           after_execution
           Utils.try(self, :"on_#{status}", e)
@@ -72,7 +72,7 @@ module Lite
           around_execution { call }
           Utils.try(self, :on_success)
         rescue StandardError => e
-          @exception = e
+          @original_exception = e
           fault(e, ERROR, metadata) unless e.is_a?(Lite::Command::Fault)
           after_execution
           Utils.try(self, :"on_#{status}", e)