+++ spec/lite/command/fault_streamer_spec.rb
@@ -24,6 +24,7 @@ RSpec.describe Lite::Command::FaultStreamer do
           expect(streamer.metadata).to eq(command.metadata)
           expect(streamer.caused_by).to be_a(EmailValidatorCommand)
           expect(streamer.thrown_by).to be_a(EmailValidatorCommand)
+          expect(streamer.fault_exception).to be_a(EmailValidatorCommand::Invalid)
         end
       end
 
@@ -38,6 +39,7 @@ RSpec.describe Lite::Command::FaultStreamer do
           expect(streamer.metadata).to eq([1, 2, 3])
           expect(streamer.caused_by).to be_a(ValidationTokenCommand)
           expect(streamer.thrown_by).to be_a(EmailValidatorCommand)
+          expect(streamer.fault_exception).to be_a(EmailValidatorCommand::Failure)
         end
       end
     end
@@ -50,6 +52,7 @@ RSpec.describe Lite::Command::FaultStreamer do
         expect(streamer.metadata).to eq(command.metadata)
         expect(streamer.caused_by).to be_a(EmailValidatorCommand)
         expect(streamer.thrown_by).to be_nil
+        expect(streamer.fault_exception).to be_nil
       end
     end
 
@@ -61,6 +64,7 @@ RSpec.describe Lite::Command::FaultStreamer do
         expect(streamer.metadata).to eq(command.metadata)
         expect(streamer.caused_by).to be_a(EmailValidatorCommand)
         expect(streamer.thrown_by).to be_nil
+        expect(streamer.fault_exception).to be_nil
       end
     end
   end