+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.1] - 2024-10-25
+### Changed
+- Add option to raise original or fault exception
+
 ## [3.1.0] - 2024-10-25
 ### Added
 - Added `raise!` method to reraise soft call errors